Implement image carousel system
Used react-bootstrap library to implement image carousel system that follows the requirments given in the prompt. The controls of the carousel as well as the navigation are hidden in the ImageCarousel.CSS file. All the images in the carousel are resized to the same dimensions in the previously mentioned CSS file.